update jquery related scripts
attempting to solve the XSS denial for thickbox, the solution is to
upgrade jquery to 1.4.4, requiring replacing the defunct thickbox with
colorbox and the old hover title with jquery-tooltip.

TODO:
 * currently the projects are grafted into this git repo, would be nicer
to use sub repositories.
 * the symlinks are ugly as well.
 * the css is not good enough (should be like old site).
 * colorbox shows image at original size - should resize to some standard (or
we could enforce standard when uploading images? but that's more work - try
to preserve old behavior).